Backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.2
argon2-cffi==25.1.0
PyJWT==2.12.1
reportlab==4.4.10
apscheduler==3.11.2
xgboost==3.2.0
# numpy 2.3+ requires Python >=3.11 (Render was on 3.10 → use PYTHON_VERSION=3.11.13).
numpy==2.4.4
joblib==1.5.3
google-cloud-firestore==2.26.0
google-cloud-secret-manager==2.21.0
firebase-admin==6.9.0
google-api-python-client==2.193.0
google-auth==2.49.1
upstash-redis==1.6.0
google-generativeai==0.8.0
groq==1.1.2
aiosqlite==0.22.0
stable-baselines3==2.8.0
scikit-learn==1.8.0
pandas==3.0.2
openpyxl==3.1.5
langgraph==1.1.6
mcp==1.27.0
celery==5.4.0
# Pin Python via Backend/runtime.txt + Render PYTHON_VERSION / .python-version (3.11.x). SB3 2.8.x requires torch>=2.3,<3.
torch==2.3.1
torch_geometric==2.5.3
# WorldMonitor data layer — added for fetcher cron service
feedparser==6.0.11
